asynchronous_main_app/requirements.txt
fastapi
websockets
azure-eventhub
uvicorn
jinja2
twelvedata
aiohttp
azure-eventhub-checkpointstoreblob-aio